+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Newpp.hs
@@ -34,8 +34,6 @@ import Cardano.Ledger.Shelley.PParams
   ( ProposedPPUpdates (..),
     emptyPPPUpdates,
   )
-import Cardano.Ledger.Shelley.TxBody (MIRPot (..))
-import Cardano.Ledger.UMapCompact (rewView)
 import Control.State.Transition
   ( STS (..),
     TRC (..),