+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/PoolReap.hs
@@ -38,7 +38,7 @@ import Cardano.Ledger.Shelley.LedgerState
   )
 import Cardano.Ledger.Shelley.TxBody (RewardAcnt, getRwdCred, ppRewardAcnt)
 import Cardano.Ledger.Slot (EpochNo (..))
-import Cardano.Ledger.UMapCompact (View (Delegations, Rewards))
+import Cardano.Ledger.UMapCompact (View (Delegations, Rewards), compactCoinOrError)
 import qualified Cardano.Ledger.UMapCompact as UM
 import Cardano.Ledger.Val ((<+>), (<->))
 import Control.SetAlgebra (dom, eval, setSingleton, (⋪), (▷), (◁))
@@ -165,7 +165,7 @@ poolReapTransition = do
       us {utxosDeposited = utxosDeposited us <-> (unclaimed <+> refunded)}
       a {asTreasury = asTreasury a <+> unclaimed}
       ( let u0 = dsUnified ds
-            u1 = (Rewards u0 UM.∪+ refunds)
+            u1 = Rewards u0 UM.∪+ Map.map compactCoinOrError refunds
             u2 = (Delegations u1 UM.⋫ retired)
          in ds {dsUnified = u2}
       )